% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\docType{data}
\name{activeMediaRules}
\alias{activeMediaRules}
\title{Currently usable breakpoints}
\format{
An object of class \code{list} of length 5.
}
\usage{
activeMediaRules
}
\description{
Currently usable breakpoints
}
\keyword{datasets}
