% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{areaBasedRules}
\alias{areaBasedRules}
\title{Generates a list of rules for a single grid attribute based on the given grid areas.}
\usage{
areaBasedRules(attribute, areas, default_value, default_rule)
}
\arguments{
\item{attribute}{The grid attribute to generate rules for}

\item{areas}{The grid areas to generate rules based on}

\item{default_value}{The default css value for the size of one element}

\item{default_rule}{Function that returns a valid css value based on the
default_value and the area details}
}
\value{
A list of attribute rules.
}
\description{
Generates a list of rules for a single grid attribute based on the given grid areas.
}
