% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseValues}
\alias{collapseValues}
\title{Takes a list of values and collapses them into a single valid css string.}
\usage{
collapseValues(values)
}
\arguments{
\item{value}{The list of values to be collapsed}
}
\value{
A valid css value string.
}
\description{
Takes a list of values and collapses them into a single valid css string.
}
