% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\docType{data}
\name{defaultMediaRules}
\alias{defaultMediaRules}
\title{Default breakpoints (Based on bootstrap specification)}
\format{
An object of class \code{list} of length 5.
}
\usage{
defaultMediaRules
}
\description{
Default breakpoints (Based on bootstrap specification)
}
\keyword{datasets}
