% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateGridRule}
\alias{generateGridRule}
\title{Generates a css rule for the given id element.}
\usage{
generateGridRule(id, attribute, value)
}
\arguments{
\item{id}{The target id of the css rule}

\item{attribute}{The css attribute the rule will target}

\item{value}{The css value for the given attribute}
}
\value{
A complete CSS rule.
}
\description{
Generates a css rule for the given id element.
}
