% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateRule}
\alias{generateRule}
\title{Generates a simple one attribute css rule for a given ID.
Mostly used to generate css grid rules.}
\usage{
generateRule(id, attribute, value)
}
\arguments{
\item{id}{CSS rule target id}

\item{attribute}{Attribute to change}

\item{value}{Value of the attribute}
}
\value{
A valid CSS string.
}
\description{
Generates a simple one attribute css rule for a given ID.
Mostly used to generate css grid rules.
}
