% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAttributeStyle}
\alias{getAttributeStyle}
\title{Generates all required css for the given css attribute based on the grid
rules provided.}
\usage{
getAttributeStyle(
  attribute,
  grid_id,
  grid_areas,
  css_target,
  css_base_value,
  css_base_rule
)
}
\arguments{
\item{attribute}{The attribute details to create rules for.}

\item{grid_id}{The id of the grid the rules will affect.}

\item{grid_areas}{The target grid area details.}

\item{css_target}{The equivalent css attribute name.}

\item{css_base_value}{The default css value for the size of one element}

\item{css_base_rule}{Function that returns a valid css value based on the
css_base_value and the area details}
}
\value{
A string of value css rules.
}
\description{
Generates all required css for the given css attribute based on the grid
rules provided.
}
