% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRuleWrapper}
\alias{getRuleWrapper}
\title{Compiles a CSS media rule template according to the given options.
Mostly used to generate media queries for diferent breakpoints.}
\usage{
getRuleWrapper(options)
}
\arguments{
\item{options}{The list of media rule options}
}
\value{
A valid CSS media rule template string.
}
\description{
Compiles a CSS media rule template according to the given options.
Mostly used to generate media queries for diferent breakpoints.
}
