% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessAttribute}
\alias{preprocessAttribute}
\title{Converts string type attributes into a named list.
Does nothing if the attribute is already a list in the correct format.}
\usage{
preprocessAttribute(attribute)
}
\arguments{
\item{attribute}{The value to process}
}
\value{
A named list.
}
\description{
Converts string type attributes into a named list.
Does nothing if the attribute is already a list in the correct format.
}
