% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocessContent}
\alias{preprocessContent}
\title{Adds a css class to any HTML elements from the content that are named.}
\usage{
preprocessContent(content)
}
\arguments{
\item{content}{A (named) list of HTML elements.}
}
\value{
A list of HTML elements.
}
\description{
Adds a css class to any HTML elements from the content that are named.
}
