% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{registerMediaRule}
\alias{registerMediaRule}
\title{Adds a media rule from the active list. Used to create cusotm breakpoints}
\usage{
registerMediaRule(name, min = NULL, max = NULL)
}
\arguments{
\item{name}{Name of the media rule to remove}

\item{min}{The minium width of the media rule (in pixels)}

\item{max}{The maximum width of the media rule (in pixels)}
}
\description{
Adds a media rule from the active list. Used to create cusotm breakpoints
}
\seealso{
[unregisterMediaRule()]

Other MediaRules: 
\code{\link{unregisterMediaRule}()}
}
\concept{MediaRules}
