% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repeatRule}
\alias{repeatRule}
\title{Used to generate css row and column value strings.
Used mostly to fill in css attributes when no argument is defined.}
\usage{
repeatRule(options)
}
\arguments{
\item{options}{the number of elements in the grid attribute}
}
\value{
A valid CSS string.
}
\description{
Used to generate css row and column value strings.
Used mostly to fill in css attributes when no argument is defined.
}
