% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniqueCols}
\alias{uniqueCols}
\title{Extracts the unique values of a area type argument.}
\usage{
uniqueCols(options)
}
\arguments{
\item{options}{The list of elements to scan}
}
\value{
A character vector.
}
\description{
Extracts the unique values of a area type argument.
}
