% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints.R
\name{unregisterMediaRule}
\alias{unregisterMediaRule}
\title{Removes a media rule from the active list.}
\usage{
unregisterMediaRule(name)
}
\arguments{
\item{name}{Name of the media rule to remove#'}
}
\description{
Removes a media rule from the active list.
}
\seealso{
[registerMediaRule()]

Other MediaRules: 
\code{\link{registerMediaRule}()}
}
\concept{MediaRules}
